Updated config-avahi script to stop if avahi-daemon is already running.
Raspbian as of 2015-05-05 release installs and runs avahi-daemon by default.

+++ bin/config-avahi.sh
@@ -3,6 +3,9 @@
 # Configure Avahi to allow ".local" host names locally.
 # Reference: http://raspberrypi.stackexchange.com/questions/7640/rpi-not-reachable-via-its-hostname-in-lan
 
+# Raspbian release as of 2015-05-05 includes avahi-daemon.
+grep -q avahi-daemon <(ps auxww | grep -v grep) && echo "*** Stopped. avahi-daemon is already running" && exit 1
+
 apt-get install avahi-daemon
 if [[ $? != 0 ]]; then
   echo "*** Failed to install avahi-daemon."
